Replaced BayesFactor models with brms
This change was made as the brms models can better account for the data structure and more closely follow our lme4 models. Additionally, I have removed the outputted models from all analyses as the brms models (saved in lists in .RData files) are too large for the 100mb capacity of GitHub. Data files will now be hosted separately on the OSF.