Fixed summariseWithin()
Fixed summariseWithin() in normalise_within_subjects_error.R to now properly handle character vector columns as grouping variables. Previously, character vectors passed as within subjects variables resulted in SDs of 0 as the function expected to be able to count the levels in the column (which is 0 when not a factor). Now uses length(unique()) to determine SD adjustment which is type insensitive.