Added final version of paper
Updated the manuscript file to include all changes to text and to include the additional appendix (included as an external .Rmd file). Also changed the exploratory plots to combine all 4 plots of error types across the two methods to allow for easy inclusion as one file in the appendix of the paper. Added a procedure table for inclusion in the paper and an additional function that is a wrapper around kable and kableExtra to avoid repetitive code.